generative_service/requirements.txt
fastapi==0.95.2
pydantic==1.10.8
uvicorn
langchain-community